Refactor some methods in Ui
This commit refactors the Ui constructor which controls the UI settings in the chatbot \n\n It serves to improve the readablity of the code in the future \n\n When there are future updates in the Ui, it will become more convenient for us to separate the ui setting for each component. In this way we have a clearer idea on what we are dealing with \n\n Let's pull up the behaviour and add more layers of abstraction to make it more readable!